pyproject.toml
[tool.poetry]
name = "fastapi-tms"
version = "0.1.0"
description = ""
authors = ["Roman <roma.makhunov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.28.0"
sqlalchemy = "^2.0.28"
alembic = "^1.13.1"
bcrypt = "^4.1.2"
aiosqlite = "^0.20.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
